move enum clases
