+++ app/src/main/java/org/eyeseetea/malariacare/domain/enums/Action.java
@@ -1,4 +1,4 @@
-package org.eyeseetea.malariacare.domain.utils;
+package org.eyeseetea.malariacare.domain.enums;
 
 public enum  Action {
     CHANGE_TAB, PRESS_BACK_BUTTON
