+++ app/src/main/java/org/eyeseetea/malariacare/domain/enums/NetworkStrategy.java
@@ -1,4 +1,4 @@
-package org.eyeseetea.malariacare.domain.utils;
+package org.eyeseetea.malariacare.domain.enums;
 
 public enum  NetworkStrategy {
     NETWORK_FIRST, LOCAL_FIRST
